+++ magenta/js/protobuf/.npmignore
