+++ magenta/js/protobuf/src/proto.d.ts
