+++ magenta/js/protobuf/src/proto.js
