+++ magenta/js/protobuf/tsconfig.json
