+++ magenta/js/protobuf/tslint.json
