+++ magenta/js/protobuf/update_proto.sh
